agent/requirements.txt
boto3
bedrock-agentcore
strands-agents[a2a]
aws_lambda_powertools
pyjwt